Move PageTable to dedicated module (#1048)
## Summary
- extract `PageTable` and `PageBatchInfo` from `attention.py`
- create new module `levanter.layers.page_table`
- update exports and imports across library and tests

## Testing
- `pre-commit run --files src/levanter/layers/page_table.py
src/levanter/layers/attention.py src/levanter/layers/__init__.py
src/levanter/models/llama.py tests/test_page_table.py
tests/test_attention.py tests/test_llama_decode.py`
- `pytest tests -m "not entry and not slow and not ray"` *(fails:
ModuleNotFoundError)*

------
https://chatgpt.com/codex/tasks/task_e_6870237bbb308331bade8b1b0966e4fe
