+++ src/levanter/layers/attention.py
@@ -25,13 +25,14 @@ from jaxtyping import PRNGKeyArray
 import haliax
 import haliax as hax
 import haliax.nn as hnn
-import haliax.haxtyping as ht
 from haliax import Axis, AxisSelection, AxisSelector, NamedArray, axis_name
 from haliax.jax_utils import maybe_rng_split, named_call
 from haliax.nn.attention import causal_mask, combine_masks_and, combine_masks_or
 from haliax.nn.normalization import LayerNormBase
 from haliax.partitioning import pspec_for_axis
 from haliax.types import PrecisionLike
+
+from .page_table import PageBatchInfo, PageTable
 
 from .normalization import LayerNormConfigBase
 from .rotary import RotaryEmbeddings, RotaryEmbeddingsConfig
@@ -1703,211 +1704,6 @@ class KvPageCache(eqx.Module):
         return KvPageCache(kv_pages)
 
 
-class PageTable(eqx.Module):
-    page_indices: NamedArray  # i32[Seq, PagePerSeq] <-- mapping from seq to pages in cache
-    page_owners: NamedArray  # i32[Page] <- int indicating which seq owns each page. -1 if free
-    seq_lens: NamedArray  # i32 <- int indicating how many tokens each seq is. -1 if unused.
-    # may be 0 if no tokens yet
-
-    page_size: int = eqx.field(static=True)
-
-    @property
-    def num_pages(self) -> int:
-        return self.page_indices.axis_size("page") * self.page_indices.axis_size("seq")
-
-    @property
-    def pages_per_seq(self) -> int:
-        return self.page_indices.axis_size("page")
-
-    @property
-    def max_seqs(self) -> int:
-        return self.page_indices.axis_size("seq")
-
-    @property
-    def max_len_per_seq(self) -> int:
-        return self.page_size * self.pages_per_seq
-
-    @property
-    def current_num_seqs(self) -> int:
-        return hax.sum(self.seq_lens >= 0).scalar()
-
-    @staticmethod
-    def init(max_pages: int, max_seqs: int, page_size: int, max_pages_per_seq: int) -> "PageTable":
-        page_indices = hax.full({"seq": max_seqs, "page": max_pages_per_seq}, -1, dtype=jnp.int32)
-        page_owners = hax.full({"page": max_pages}, -1, dtype=jnp.int32)
-        seq_lens = hax.full({"seq": max_seqs}, -1, dtype=jnp.int32)
-        return PageTable(page_indices, page_owners, seq_lens, page_size)
-
-    @eqx.filter_jit(donate="all")
-    def assign_seq_id_to_seq(self) -> tuple["PageTable", int]:
-        # find the first -1 in seq_lens
-        seq_id = hax.argmin(self.seq_lens, "seq").scalar()
-        # seq_id = eqx.error_if(seq_id, self.seq_lens["seq", seq_id] != -1, "No unused seqs")
-        new_seq_lens = self.seq_lens.at["seq", seq_id].set(0)
-
-        return dataclasses.replace(self, seq_lens=new_seq_lens), seq_id
-
-    # @eqx.filter_jit(donate="all")
-    def allocate_for_seqs(
-        self,
-        updated_seqs: ht.i32[NamedArray, " seq"],  # type: ignore[name-defined]
-        new_counts: ht.i32[NamedArray, " seq"],  # type: ignore[name-defined]
-        tokens: ht.i32[NamedArray, "position"],  # type: ignore[name-defined]
-    ) -> tuple["PageTable", "PageBatchInfo"]:
-        """
-        Allocate pages for new sequences and updates seq_lens
-
-        Args:
-            updated_seqs: i32[Seq] ids for the updated sequences (can be padded with -1s). Can be a
-                NamedArray or a plain ndarray.
-            new_counts: NamedArray i32[Seq] number of new tokens for each sequence
-            tokens: NamedArray i32[position] sequence id for each new token. Values
-                should be padded with -1
-
-        Returns:
-            PageTable with new seq_lens and allocated pages, PageBatchInfo with page indices and seq_lens
-        """
-        # convert NamedArrays to plain arrays for easier manipulation inside jit/loops
-        page_indices = self.page_indices
-        page_owners = self.page_owners
-        seq_lens = self.seq_lens
-
-        padded_updated_seqs = hax.where(updated_seqs < 0, self.max_seqs, updated_seqs)
-        current_lens = hax.where(seq_lens < 0, 0, seq_lens)
-        new_lens_tmp = current_lens.at["seq", padded_updated_seqs].add(new_counts, mode="drop")
-        new_lens = hax.where(seq_lens < 0, hax.where(new_lens_tmp > 0, new_lens_tmp, -1), new_lens_tmp)
-
-        new_num_pages_needed = (new_lens + self.page_size - 1) // self.page_size
-        old_num_pages_needed = (seq_lens + self.page_size - 1) // self.page_size
-
-        def _alloc_pages_for_seq(seq_id, carry):
-            page_indices, page_owners = carry
-            num_needed = new_num_pages_needed["seq", seq_id].scalar()
-            old_needed = old_num_pages_needed["seq", seq_id].scalar()
-
-            def body(page_idx, state):
-                page_indices, page_owners = state
-                free_page_idx = hax.argmin(page_owners, "page")
-                # free_page_idx = eqx.error_if(free_page_idx, page_owners["page", free_page_idx] != -1, "No free pages")
-                page_owners = page_owners.at["page", free_page_idx].set(seq_id)
-                page_indices = page_indices.at["seq", seq_id, "page", page_idx].set(free_page_idx)
-                return page_indices, page_owners
-
-            new_page_indices, new_page_owners = jax.lax.fori_loop(
-                old_needed, num_needed, body, (page_indices, page_owners)
-            )
-            return new_page_indices, new_page_owners
-
-        page_indices, page_owners = jax.lax.fori_loop(
-            0, self.max_seqs, _alloc_pages_for_seq, (page_indices, page_owners)
-        )
-
-        new_table = dataclasses.replace(
-            self,
-            page_indices=page_indices,
-            page_owners=page_owners,
-            seq_lens=new_lens,
-        )
-
-        batch_info = self._slice_batch_info(updated_seqs, self.seq_lens, new_table, new_counts, tokens)
-
-        return new_table, batch_info
-
-    def _slice_batch_info(self, updated_seqs, old_seq_lens, new_table, new_token_counts, tokens):
-        mask = updated_seqs >= 0
-        safe_updated = hax.where(mask, updated_seqs, 0)
-
-        gathered_page_indices = new_table.page_indices["seq", safe_updated]
-        page_indices = hax.where(mask, gathered_page_indices, -1)
-
-        seq_lens = new_table.seq_lens["seq", safe_updated]
-        seq_lens = hax.where(mask, seq_lens, -1)
-
-        num_seqs = hax.sum(mask).scalar()
-
-        # Initialize destination indices as a NamedArray instead of a raw JAX array
-        token_dests = hax.full(tokens.shape, -1, dtype=jnp.int32)
-
-        # start offsets for each sequence (treat -1 lens as 0)
-        seq_cursors = jnp.where(old_seq_lens.array < 0, 0, old_seq_lens.array)
-
-        def token_body(i, carry):
-            token_dests, seq_cursors = carry
-            seq_id = tokens["position", i].scalar()
-
-            def assign(carry):
-                token_dests, seq_cursors = carry
-                page_idx = seq_cursors[seq_id] // self.page_size
-                page_offset = seq_cursors[seq_id] % self.page_size
-                page = new_table.page_indices["seq", seq_id, "page", page_idx]
-                dest = hax.where(page < 0, -1, page * self.page_size + page_offset)
-                token_dests = token_dests.at["position", i].set(dest)
-                seq_cursors = seq_cursors.at[seq_id].add(1)
-                return token_dests, seq_cursors
-
-            token_dests, seq_cursors = jax.lax.cond(seq_id >= 0, assign, lambda c: c, (token_dests, seq_cursors))
-            return token_dests, seq_cursors
-
-        token_dests, _ = jax.lax.fori_loop(0, tokens.axis_size("position"), token_body, (token_dests, seq_cursors))
-
-        cu_q_lens = hax.concatenate(
-            "seq",
-            [
-                hax.zeros({"seq": 1}, dtype=jnp.int32),  # cu_q_lens[0] = 0
-                hax.cumsum(new_token_counts, "seq", dtype=jnp.int32),
-            ],
-        )
-        batch_info = PageBatchInfo(
-            page_indices=page_indices,
-            seq_lens=seq_lens,
-            cu_q_lens=cu_q_lens,
-            num_seqs=num_seqs,
-            new_token_dests=token_dests,
-            page_size=self.page_size,
-        )
-        return batch_info
-
-    @eqx.filter_jit(donate="all")
-    def free_pages(self, seq_id: int) -> "PageTable":
-        # find all pages owned by this seq
-
-        # pages_owned = self.page_owners == seq_id
-        # new_page_owners = self.page_owners.at["page", pages_owned].set(-1)
-        new_page_owners = hax.where(self.page_owners == seq_id, -1, self.page_owners)
-        new_page_indices = self.page_indices.at["seq", seq_id].set(-1)
-        new_seq_lens = self.seq_lens.at["seq", seq_id].set(-1)
-
-        return dataclasses.replace(
-            self,
-            page_owners=new_page_owners,
-            page_indices=new_page_indices,
-            seq_lens=new_seq_lens,
-        )
-
-
-class PageBatchInfo(eqx.Module):
-    """
-    Contains just the page and length information for a particular set of sequences, -1 padded.
-
-    Arguments:
-        page_indices: i32[Seq, Page] mapping from seq to pages in cache
-        seq_lens: i32[Seq]. Contains the lengths **after** the new tokens are appended.
-        cu_q_lens: i32[Seq + 1]. Cumulative lengths for the sequences, including new tokens
-        num_seqs: i32[] number of sequences in the batch
-        page_size: int
-    """
-
-    # NOTE: seq_lens is the length of the sequence **after** the new tokens are appended, though these
-    # positions won't have been filled in the kv cache typically
-    page_indices: ht.i32[NamedArray, " seq page"]
-    seq_lens: ht.i32[NamedArray, " seq"]
-    cu_q_lens: ht.i32[NamedArray, " seq"]  # cumulative lengths for the sequences, including new tokens
-    num_seqs: ht.i32[jnp.ndarray, ""]  # number of sequences in the batch
-    new_token_dests: ht.i32[NamedArray, "position"]  # ipage indices for the new tokens, -1 padded
-    page_size: int = eqx.field(static=True)
-
-    def __post_init__(self):
-        assert isinstance(self.num_seqs, jnp.ndarray), "num_seqs must be a JAX ndarray"
 
 
 class KvPageState(eqx.Module):
